requirements.txt
requests
python-dotenv
flask
flask-sock
gevent
gevent-websocket
typer
